Move SearchContext to contexts/displays
